Move test helpers to a different file
